Correct FoxClient destruction of socket.

+++ Fox/src/FoxClient.cpp
@@ -109,12 +109,7 @@ FoxClient::~FoxClient()
       delete m_sendingTimer;
    }
 
-   if(mpClient!=0) {
-      
-      mpClient->Destroy();
-      delete mpClient;
-      mpClient = 0;
-   }
+   if(mpClient!=0) mpClient->Destroy();
    delete m_DataMutex;
 }
 void FoxClient::setNbOfAvailCPUs(int nb)
